+++ huey/tests/__init__.py
@@ -2,6 +2,7 @@ from huey.tests.test_api import *
 from huey.tests.test_asyncio import *
 from huey.tests.test_consumer import *
 from huey.tests.test_crontab import *
+from huey.tests.test_django_outbox import *
 from huey.tests.test_django_tasks import *
 from huey.tests.test_helpers import *
 from huey.tests.test_immediate import *